backend/pyproject.toml
[project]
name = "spatial-sound-anomaly"
version = "0.1.0"
description = "Spatial sound anomaly detection system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "numpy>=1.26",
    "httpx>=0.27",
    "sentencepiece>=0.2",
    "accelerate>=0.27",
]

[project.optional-dependencies]
gpu = [
    "torch>=2.0",
    "torchaudio>=2.0",
    "transformers>=4.40",
    # faiss-gpu has no Python 3.11+ wheels; use faiss-cpu + GPU resources at runtime
    "faiss-cpu>=1.7",
    "librosa>=0.10",
    "datasets>=3.0",
    "soundfile>=0.12",
]
cpu = [
    "torch>=2.0",
    "torchaudio>=2.0",
    "transformers>=4.40",
    "faiss-cpu>=1.7",
    "librosa>=0.10",
    "datasets>=3.0",
    "soundfile>=0.12",
]
dev = [
    "ruff>=0.8",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
